requirements.txt
# Video Translator Pro - Python Dependencies

# Core AI/ML Libraries
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Translation
deep-translator>=1.11.4

# GUI (built-in with Python, no install needed)
# tkinter - comes with Python

# System dependencies (install separately):
# - FFmpeg: https://ffmpeg.org/download.html
#   Windows: Download from https://www.gyan.dev/ffmpeg/builds/
#   macOS: brew install ffmpeg
#   Linux: sudo apt-get install ffmpeg

# Optional: Faster CPU inference
# numpy>=1.24.0

# Development Dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0